backend/requirements/base.txt
Django==5.2
djangorestframework==3.17.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
psycopg2-binary==2.9.11
Pillow==12.2.0
